requirements.txt
git+https://github.com/Rapptz/discord.py.git
git+https://github.com/Tom-the-Bomb/Discord-Games.git
git+https://github.com/f-stop-lang/f-stop-rply.git
typing_extensions
matplotlib
humanize
jishaku
Pillow
wand
psutil
sympy
opencv-python
async_tio
pilmoji
glitch-this